requirements.txt
PyQt5
numpy
cupy-cuda11x  # Replace '11x' with your CUDA version, or omit if not using CuPy
matplotlib
scipy
opencv-python
scikit-image
psutil